+++ RomiboController/LXReorderableCollectionViewFlowLayout.m
@@ -215,7 +215,7 @@ static NSString * const kLXCollectionViewKeyPath = @"collectionView";
     CGPoint contentOffset = self.collectionView.contentOffset;
     // Important to have an integer `distance` as the `contentOffset` property automatically gets rounded
     // and it would diverge from the view's center resulting in a "cell is slipping away under finger"-bug.
-    CGFloat distance = rint(self.scrollingSpeed / LX_FRAMES_PER_SECOND);
+    CGFloat distance = (float)rint(self.scrollingSpeed / LX_FRAMES_PER_SECOND);
     CGPoint translation = CGPointZero;
     
     switch(direction) {
